+++ tests/test_pipeline.py
@@ -1,6 +1,8 @@
+import pytest
 import trio
 
 from slurry import Pipeline
+from slurry.sections import Map
 from slurry.sections.abc import Section
 
 from .fixtures import produce_increasing_integers
@@ -44,3 +46,21 @@ async def test_early_tap_closure_section(autojump_clock):
             async for i in aiter:
                 assert isinstance(i, int)
                 break
+
+async def test_welding(autojump_clock):
+    async with Pipeline.create(
+        produce_increasing_integers(1),
+        (Map(lambda i: i+1),)
+        ) as pipeline:
+        async with pipeline.tap() as aiter:
+            result = [i async for i in aiter]
+            assert result == [1, 2, 3]
+
+async def test_welding_two_generator_functions_not_allowed(autojump_clock):
+    with pytest.raises(ValueError):
+        async with Pipeline.create(
+            produce_increasing_integers(1),
+            produce_increasing_integers(1),
+        ) as pipeline, pipeline.tap() as aiter:
+            result = [i async for i in aiter]
+